#1067 Fix as the UtilDateTime.getDayEnd slightly changed
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@12933 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/financials/src/com/opensourcestrategies/financials/invoice/InvoiceServices.java
@@ -1268,7 +1268,7 @@ public final class InvoiceServices {
                 cal.set(Calendar.DAY_OF_MONTH, days.intValue());
                 dueDate = UtilDateTime.getDayEnd(new Timestamp(cal.getTime().getTime()), timeZone, locale);
             } else if (UtilValidate.isNotEmpty(days)) {
-                dueDate = UtilDateTime.getDayEnd(UtilDateTime.nowTimestamp(), days.intValue(), timeZone, locale);
+                dueDate = UtilDateTime.getDayEnd(UtilDateTime.nowTimestamp(), days, timeZone, locale);
             }
 
             if (UtilValidate.isNotEmpty(dueDate)) {
@@ -2510,7 +2510,7 @@ return results;
             Long orderTermNetDays = orh.getOrderTermNetDays();
             Timestamp dueDate = null;
             if (orderTermNetDays != null) {
-                dueDate = UtilDateTime.getDayEnd(invoiceDate, orderTermNetDays.intValue());
+                dueDate = UtilDateTime.getDayEnd(invoiceDate, orderTermNetDays);
             }
 
             // create the invoice record
